Fixes #78 Fill the fact sheet (#79)
* WIP

* images dummy

* Update factsheet.md

* Add files via upload

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md (#2)

* Update factsheet.md (#3)

* Update factsheet.md

Could you adapt the side width? The slider for the horizontal axis is at the bottom of the page and thus not handy to see how a sentence continues (e.g. line 39)
Line 13: Some of the special populations mentioned need to be incorporated manually (HI, RI, Obese)
Why do we have a dedicated section on "Permeabilits" (line 114) although this is already covered in the compound section?
blabla (line 257)?
Features of MoBi could be listed in the same tabular style like for PK-Sim

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Update factsheet.md

* Delete dummy.jpg

* Delete image001.png

* Delete image003.png

* Delete image005.png

* Delete image007.png

* Delete image009.png

* Delete image011.png

* Delete image013.png

* Delete image015.png

* Delete image017.png

* Delete image023.png

* Update factsheet.md